refactor(settings): extract shared chips; clarify review nits
Move the reusable SpaceChip/WrapChips out of the #6 feature file
into a neutral Chips.swift so #7 no longer depends on a #6 file for
shared primitives. Add clarifying comments on the "separate Spaces"
absent-value default and the rename keybinding-arg scope.
