+++ Sources/KiwiDesk/Settings/Tabs/SpaceMonitorAssignments.swift
@@ -97,48 +97,3 @@ struct SpaceMonitorAssignments: View {
         let monitor: String
     }
 }
-
-/// A pill rendering a space name, shared by the palette and the
-/// assignment rows.
-struct SpaceChip: View {
-    let label: String
-
-    var body: some View {
-        Text(label)
-            .font(.caption)
-            .fontWeight(.medium)
-            .padding(.horizontal, 8)
-            .padding(.vertical, 3)
-            .background(
-                Capsule().fill(.tint.opacity(0.15))
-            )
-            .overlay(
-                Capsule().strokeBorder(.tint.opacity(0.4))
-            )
-    }
-}
-
-/// A simple left-to-right flowing row of chips that wraps to the
-/// next line when it runs out of width.
-struct WrapChips<Item, Chip: View>: View {
-    private let items: [Item]
-    private let chip: (Item) -> Chip
-
-    init(
-        _ items: [Item],
-        @ViewBuilder chip: @escaping (Item) -> Chip
-    ) {
-        self.items = items
-        self.chip = chip
-    }
-
-    var body: some View {
-        FlowLayout(spacing: 6) {
-            ForEach(Array(items.enumerated()), id: \.offset) {
-                _,
-                item in
-                chip(item)
-            }
-        }
-    }
-}
